+++ configs.py
@@ -43,12 +43,15 @@ args.multi_gpus = torch.cuda.device_count() > 0
 
 args.seq_len_ques       = 40
 args.seq_len_para       = 40
-args.seq_len_ans        = 12
+args.seq_len_ans        = 20
 args.n_paras            = 40
-args.dim_hid            = 256
+args.d_embd             = 200
+args.d_hid              = 256
 args.max_len_ans        = 12
-args.teacher_forcing    = 0.3
 args.min_count_PGD      = 10        # min occurences of word to be added to vocab of PointerGeneratorDecoder
+args.d_vocab            = 131059    # Vocab size
+args.dropout            = 0.2
+args.n_layers           = 5
 
 
 ###############################
@@ -62,6 +65,8 @@ logging.getLogger().setLevel(logging.INFO)
 ###############################
 PATH    = {
     ## Paths associated with Data Reading
-    'dataset_para'  : "./backup/[SPLIT]/data_[SHARD].csv",
-    'vocab_PGD'     : "./backup/vocab_PGD.txt"
+    'dataset_para'  : "backup/[SPLIT]/data_[SHARD].csv",
+    'vocab_PGD'     : "backup/vocab_PGD.txt",
+    'saved_model'   : "backup/model.pt",
+    'saved_chkpoint': "backup/chkpoint.pth.tar"
 }