+++ launcher.sh
@@ -6,11 +6,13 @@
 
 launch()
 {
+    mkdir -p backup
+
     # echo "1. Decompose each document in dataset into paragraphs."
-    python -m modules.data_reading.data_reading --num_proc 5
+    # python -m modules.data_reading.data_reading --num_proc 5
 
     # echo "2. Start training"
-    python -m modules.narrativepipeline.NarrativePipeline
+    python -m modules.narrativepipeline.NarrativePipeline --batch 5
 }