+++ modules/Reasoning/IAL.py
@@ -2,51 +2,42 @@ import torch.nn.functional as torch_f
 import torch.nn as torch_nn
 import torch
 
+from modules.utils import NonLinear, transpose
 from configs import args
 
 Block   = 200
 
 class IntrospectiveAlignmentLayer(torch_nn.Module):
-    def __init__(self, d_hid=args.dim_hid, d_emb=200):
+    def __init__(self, d_hid=args.d_hid, d_emb=200):
         super().__init__()
 
         self.d_hid  = d_hid
-        self.d_emb  = d_emb
 
 
         self.biLSTM_emb     = torch_nn.LSTM(d_emb, d_hid//2, num_layers=5,
                                            batch_first=True, bidirectional=True)
 
-        self.linearIAL      = torch_nn.Linear(d_hid, d_hid)
+        self.linearIAL      = NonLinear(d_hid, d_hid)
 
         self.linearReason   = torch_nn.Linear(4*d_hid, 4*d_hid)
 
         self.biLSTM_attn    = torch_nn.LSTM(8*d_hid, d_hid, num_layers=5,
                                            batch_first=True, bidirectional=True)
 
-    def forward(self, ques, paras):
-        # ques       : [batch, seq_len_ques, d_embd]
-        # paras      : [batch, seq_len_contex, d_embd]
+    def forward(self, H_q, H_c):
+        # H_q   : [batch, seq_len_ques, d_hid]
+        # H_c   : [batch, seq_len_contex, d_hid]
 
-        batch, seq_len_context, _ = paras.shape
-
-        ques    = torch_f.relu(ques)
-        paras   = torch_f.relu(paras)
-
-        # Input and Context embedding
-        H_q = self.biLSTM_emb(ques)[0]
-        H_c = self.biLSTM_emb(paras)[0]
-        # H_q: [batch, seq_len_ques, d_hid]
-        # H_c: [batch, seq_len_context, d_hid]
+        batch, seq_len_context, _ = H_c.shape
 
 
         # Introspective Alignment
-        H_q = torch.sigmoid(self.linearIAL(H_q))
-        H_c = torch.sigmoid(self.linearIAL(H_c))
+        H_q = self.linearIAL(H_q)
+        H_c = self.linearIAL(H_c)
         # H_q: [batch, seq_len_ques, d_hid]
         # H_c: [batch, seq_len_context, d_hid]
 
-        E   = torch.bmm(H_c, torch.reshape(H_q, (batch, self.d_hid, -1)))
+        E   = torch.bmm(H_c, transpose(H_q))
         # E: [batch, seq_len_context, seq_len_ques]
         A   = torch.bmm(torch_f.softmax(E, dim=1), H_q)
         # A: [batch, seq_len_context, d_hid]